Moving the addresses, status, and email array into the 'extra' portion of the auth hash to be more compliant with the omniauth shema.
Changing the uid of the OAuth from the client id to the first returned email of the account. This is the best I can do for a UID I believe at this time. The previous use of client.id can cause conflicts in finding tokens based on provider and uid if two accounts are used on the same computer.